chore: create useIconProps
