--- packages/blade/src/utils/getIconDimensions/getIconDimensions.test.ts
@@ -1,1 +1,0 @@
-// TODO
