--- packages/blade/src/utils/getIconDimensions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './getIconDimensions';
